Refactor retrieving information from game object
