+++ api/aiagent-orchestrator.py
@@ -3,17 +3,12 @@ import json
 import re
 import logging
 import os
+from shared.game import get_player_by_nickname
 logger = logging.getLogger()
 logger.setLevel(logging.INFO)
 
 
 lambda_client = boto3.client('lambda')
-
-
-def get_player_by_nickname(players, nickname):
-    filtered_players = [p for p in players if p["nickname"] == nickname]
-    if filtered_players:
-        return filtered_players[0]
 
 
 def get_aiagent_name(game):
